Use archive.debian.org for apt sources on debian 7
The debian 7 (wheezy) mirrors were archived since it's EOL.
I don't think we statically link with anything from wheezy we just use it because it
has an early version of glibc so we get decent OS compatibility. But we should
still consider moving to a new version.

https://lists.debian.org/debian-devel-announce/2019/03/msg00006.html

I added templating so I could make the inclusion of the sources.list file conditional.