fix(go/base): set GOTOOLCHAIN=local and skip building arm using the base Dockerfile (#524)
Disable the automatic downloads of the Go toolchain.

For the goal of reproducible builds using golang-crossbuild, we want all of the binaries used to produce our artifacts to be self-contained in the image. Adding GOTOOLCHAIN=local disable downloads of newer toolchains.

This changes the entrypoint of the images to document that GOTOOLCHAIN is set during builds.  And it sets GOTOOLCHAIN=local in the Dockerfile so that when building entrypoint.go it uses the toolchain that was previously installed.

The two Dockerfiles (base and base-arm) had different 'go build' commands. This unifies them to both have CGO_ENABLED=0 and to explicitly set GOARCH to avoid accidentally producing a binary for the wrong target architecture (which we've seen happen for unknown reasons).

Additionally we remove the arm target from the base Dockerfile, to address the CI concurrency problem described
in https://github.com/elastic/golang-crossbuild/pull/523#issuecomment-2688537270

https://go.dev/doc/toolchain

Co-Authored by @andrewkroh